ccps-project/server/requirements.txt
absl-py==2.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
flash==1.0.3
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.2.0
grpcio==1.62.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
protobuf==5.26.1
psutil==5.9.8
pvporcupine==3.0.2
pvrecorder==1.2.2
py-cpuinfo==9.0.0
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.3.0+cpu
torchaudio==2.3.0+cpu
torchvision==0.18.0+cpu
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.2
urllib3==2.2.1
Werkzeug==3.0.2
